Extract formatting tasks from task.py
