+++ exasol/toolbox/nox/tasks.py
@@ -29,6 +29,11 @@ from exasol.toolbox.metrics import (
     create_report,
     format_report,
 )
+from exasol.toolbox.nox._format import (
+    _code_format,
+    _pyupgrade,
+    fix,
+)
 from exasol.toolbox.nox._release import prepare_release
 from exasol.toolbox.nox._shared import (
     DOCS_OUTPUT_DIR,
@@ -49,26 +54,6 @@ from noxconfig import (
 )
 
 
-def _code_format(session: Session, mode: Mode, files: Iterable[str]) -> None:
-    isort = ["poetry", "run", "isort", "-v"]
-    black = ["poetry", "run", "black"]
-    isort = isort if mode == Mode.Fix else isort + ["--check"]
-    black = black if mode == Mode.Fix else black + ["--check"]
-    session.run(*isort, *files)
-    session.run(*black, *files)
-
-
-def _pyupgrade(session: Session, files: Iterable[str]) -> None:
-    session.run(
-        "poetry",
-        "run",
-        "pyupgrade",
-        "--py38-plus",
-        "--exit-zero-even-if-changed",
-        *files,
-    )
-
-
 def _pylint(session: Session, files: Iterable[str]) -> None:
     session.run("poetry", "run", "python", "-m", "pylint", *files)
 
@@ -87,15 +72,6 @@ def _type_check(session: Session, files: Iterable[str]) -> None:
         "--scripts-are-modules",
         *files,
     )
-
-
-@nox.session(python=False)
-def fix(session: Session) -> None:
-    """Runs all automated fixes on the code base"""
-    py_files = [f"{file}" for file in python_files(PROJECT_CONFIG.root)]
-    _version(session, Mode.Fix, PROJECT_CONFIG.version_file)
-    _pyupgrade(session, py_files)
-    _code_format(session, Mode.Fix, py_files)
 
 
 @nox.session(name="check", python=False)
